refactor: move collections related module in a collections package
