+++ tdp/core/collections/collections.py
@@ -17,12 +17,13 @@ from collections.abc import Iterable
 from pathlib import Path
 from typing import TYPE_CHECKING, Optional
 
-from tdp.core.collection_reader import CollectionReader
 from tdp.core.entities.hostable_entity_name import ServiceComponentName
 from tdp.core.entities.operation import Operations, Playbook
 from tdp.core.inventory_reader import InventoryReader
 from tdp.core.operation import Operation
 from tdp.core.variables.schema.service_schema import ServiceSchema
+
+from .collection_reader import CollectionReader
 
 if TYPE_CHECKING:
     from tdp.core.types import PathLike
