+++ test_dag_order/helpers.py
@@ -10,13 +10,13 @@ from typing import Any, TextIO, Union
 
 import yaml
 
-from tdp.core.collection_reader import (
+from tdp.core.collections import Collections
+from tdp.core.collections.collection_reader import (
     CollectionReader,
     MissingMandatoryDirectoryError,
     PathDoesNotExistsError,
     PathIsNotADirectoryError,
 )
-from tdp.core.collections import Collections
 from tdp.core.constants import YML_EXTENSION
 from tdp.core.entities.hostable_entity_name import (
     ServiceName,
